changes on unity

+++ scripts/eulerian_case3.py
@@ -170,14 +170,16 @@ else:
 C0 = pdf_IC(params.z_cell)[None,:] * params.mass_fractions[:,None]
 
 datafolder = '/work6/torn/EulerLagrange'
-datafolder = '../results/'
+#datafolder = '../results/'
 outputfilename = os.path.join(datafolder, f'Case3_K_{label}_block_Nclasses={params.Nclasses}_NJ={params.Nz}_dt={params.dt}.npy')
 
+if os.path.exists(outputfilename):
+    print('File exists, exiting: ', outputfilename)
+else:
+    tic = time.time()
+    c = Crank_Nicolson_FVM_TVD_advection_diffusion_reaction(C0, K, params, outputfilename = outputfilename)
+    toc = time.time()
+    print(f'Simulation took {toc - tic:.1f} seconds, output written to {outputfilename}')
 
-tic = time.time()
-c = Crank_Nicolson_FVM_TVD_advection_diffusion_reaction(C0, K, params, outputfilename = outputfilename)
-toc = time.time()
-print(f'Simulation took {toc - tic:.1f} seconds, output written to {outputfilename}')
-
-np.save(outputfilename, c)
+    np.save(outputfilename, c)
